API & UI - Configured Docker containerization

+++ backend/TaskInsightEngine/TaskInsightEngine.Application/Dtos/Risk/RiskBaselineDto.cs
@@ -2,7 +2,7 @@
 {
     public class RiskBaselineDto
     {
-        public long TaskId { get; set; }
+        public int TaskId { get; set; }
         public int CurrentScore { get; set; }
         public string? CurrentLevel { get; set; }
     }
